SOLID원칙을 참고하여 함수로직 변경 및 재활용함수 분리
