+++ weekly-mission/login.js
@@ -1,28 +1,18 @@
-const emailInput = document.querySelector('#email');
+import {returnInputValue, addErrorClass, eyeIconClickEvent} from './recycle.js'
 const errorMessageEmail = document.querySelector('.error-message[data-about=email]');
 const passwordInput = document.querySelector('#password');
-const eyeIconImg = document.querySelector('.show-password > img');
-const signinForm = document.querySelector('#form-signin');
+
 
 //이메일 형식 체크
-function checkEmailInput(){
+function checkEmailInput(event){
+    const userInput = returnInputValue(event);
+    console.log(userInput)
     let message = emailInput.value === '' ? '이메일을 입력해주세요.' : /@/g.test(emailInput.value) === false ? '올바른 이메일 형식이 아닙니다.' : '';
     errorMessageEmail.textContent = message;
-    message !== '' ?  emailInput.classList.add('error') : emailInput.classList.remove('error');
-
+    addErrorClass(event.target, message);
 }
 
-//비밀번호 표시 아이콘 이벤트
-function eyeIconClickEvent(e){
-    if(e.target.tagName === 'IMG'){
-        e.target.classList.toggle('hidden')
-        const input = e.target.parentNode.previousElementSibling;
-        const imgSrc = e.target.classList.contains('hidden') ? './img/eye-off.svg' : './img/eye-on.svg';
-        e.target.src = imgSrc;
-        e.target.classList.contains('hidden') ? input.type = 'password' : input.type = 'text';
-        return;
-    }
-}
+
 
 //폼 제출 이벤트
 function submitFormEvent(e){
@@ -34,10 +24,10 @@ function submitFormEvent(e){
     } 
 }
 
-emailInput.addEventListener('focusout', checkEmailInput);
+document.querySelector('#email').addEventListener('focusout', checkEmailInput);
 
-signinForm.addEventListener('submit',submitFormEvent );
+document.querySelector('#form-signin').addEventListener('submit',submitFormEvent );
 
-eyeIconImg.addEventListener('click',eyeIconClickEvent );
+document.querySelector('.show-password > img').addEventListener('click',eyeIconClickEvent );
