+++ weekly-mission/signin.html
@@ -58,7 +58,7 @@
         </div>
     </div>
   
-    <script src="./login.js"></script>
+    <script type="module" src="./login.js"></script>
 </body>
 
 </html>
