+++ weekly-mission/signup.html
@@ -66,6 +66,6 @@
         </div>
     </div>
 
-    <script src="./signup.js"></script>
+    <script type="module" src="./signup.js"></script>
 </body>
 </html>
